move generic writers
