+++ swiftwinrt/code_writers/writer_helpers.h
@@ -456,42 +456,6 @@ bind<write_abi_args>(function));
     static void write_implementable_interface(writer& w, interface_type const& type);
 
 
-    static void write_ireference_init_extension(writer& w, generic_inst const& type)
-    {
-        if (!is_winrt_ireference(type)) return;
-        auto generic_param = type.generic_params()[0];
-        w.add_depends(*generic_param);
-
-        auto impl_names = w.push_impl_names(true);
-
-        w.write(R"(internal enum %: ReferenceBridge {
-    typealias CABI = %
-    typealias SwiftProjection = %
-    static var IID: %.IID { IID_% }
-
-    static func from(abi: ComPtr<CABI>?) -> SwiftProjection? {
-        guard let val = abi else { return nil }
-        var result: %%
-        try! CHECKED(val.get().pointee.lpVtbl.pointee.get_Value(val.get(), &result))
-        return %
-    }
-
-    static func makeAbi() -> CABI {
-        let vtblPtr = withUnsafeMutablePointer(to: &%VTable) { $0 }
-        return .init(lpVtbl: vtblPtr)
-    }
-}
-)", bind_bridge_name(type),
-    type.mangled_name(),
-    get_full_swift_type_name(w, generic_param),
-    w.support,
-    type.mangled_name(),
-    bind<write_type>(*generic_param, write_type_params::c_abi),
-    bind<write_default_init_assignment>(*generic_param, projection_layer::c_abi),
-    bind<write_consume_type>(generic_param, "result", true),
-    type.mangled_name());
-    }
-
     static void write_class_func_body(writer& w, function_def const& function, interface_info const& iface, bool is_noexcept);
     static void write_comma_param_names(writer& w, std::vector<function_param> const& params);
     template <typename T>
@@ -529,14 +493,6 @@ typealias % = InterfaceWrapperBase<%>
 )";
         w.write(format, wrapper_name, impl_name);
     }
-
-    static void write_generic_delegate_wrapper(writer& w, generic_inst const& generic)
-    {
-        write_delegate_wrapper(w, generic);
-    }
-
-
-
 
     static void write_convert_vtable_params(writer& w, function_def const& signature)
     {
@@ -1079,59 +1035,6 @@ bind_bridge_fullname(type));
 
 
     void write_interface_bridge(writer& w, metadata_type const& type);
-
-    static void write_generic_interface_implementation(writer& w, generic_inst const& type)
-    {
-        write_interface_bridge(w, type);
-
-        w.write("fileprivate class % : %, AbiInterfaceImpl {\n",
-            bind_impl_name(type), type.generic_type_abi_name());
-
-        auto indent_guard = w.push_indent();
-
-        write_generic_typealiases(w, type);
-
-
-        w.write("typealias Bridge = %\n", bind_bridge_name(type));
-        w.write("let _default: Bridge.SwiftABI\n");
-
-        w.write("init(_ fromAbi: ComPtr<Bridge.CABI>) {\n");
-        w.write("    _default = Bridge.SwiftABI(fromAbi)\n");
-        w.write("}\n\n");
-
-        interface_info info{ &type };
-        info.is_default = true; // mark as default so we use the name "_default"
-
-        // Due to https://linear.app/the-browser-company/issue/WIN-148/investigate-possible-compiler-bug-crash-when-generating-collection
-        // we have to generate the protocol conformance for the Collection protocol (see "// MARK: Collection" below). We shouldn't have to
-        // do this because we define an extension on the protocol which does this.
-        write_collection_protocol_conformance(w, info);
-
-        for (const auto& method : type.functions)
-        {
-            write_class_impl_func(w, method, info, *type.generic_type());
-        }
-        for (const auto& prop : type.properties)
-        {
-            write_class_impl_property(w, prop, info, *type.generic_type());
-        }
-        for (const auto& event : type.events)
-        {
-            write_class_impl_event(w, event, info, *type.generic_type());
-        }
-
-        for (const auto& [interface_name, info] : type.required_interfaces)
-        {
-            if (!can_write(w, info.type)) { continue; }
-
-            write_interface_impl_members(w, info, *type.generic_type());
-        }
-
-        w.write("public func queryInterface(_ iid: %.IID) -> IUnknownRef? { nil }\n", w.support);
-        indent_guard.end();
-        w.write("}\n\n");
-    }
-
 
     static void write_param_names(writer& w, std::vector<function_param> const& params, std::string_view format)
     {
@@ -2017,11 +1920,6 @@ public init<Composable: ComposableImpl>(
     return S_OK
 })"
 );
-    }
-
-    static void write_iinspectable_methods(writer& w, generic_inst const& type)
-    {
-        write_iinspectable_methods(w, type, type.required_interfaces);
     }
 
     // assigns return or out parameters in vtable methods
