+++ swiftwinrt/types/mapped_type.h
@@ -68,10 +68,6 @@ namespace swiftwinrt
 
         void write_c_abi_param(writer& w) const override;
 
-        void write_swift_declaration(writer&) const override
-        {
-        }
-
         bool is_experimental() const override
         {
             return false;
